sonogram moved to sonogram, curio monitor in sep window
